feat : image 조회 삭제 추가 + 글과 분리
