+++ src/main/java/com/sparta/restplaceforj/controller/PostController.java
@@ -3,8 +3,6 @@ package com.sparta.restplaceforj.controller;
 import com.sparta.restplaceforj.common.CommonResponse;
 import com.sparta.restplaceforj.common.ResponseEnum;
 import com.sparta.restplaceforj.dto.AddCardRequestDto;
-import com.sparta.restplaceforj.dto.CardRequestDto;
-import com.sparta.restplaceforj.dto.ImageResponseDto;
 import com.sparta.restplaceforj.dto.PageResponseDto;
 import com.sparta.restplaceforj.dto.PostIdTitleDto;
 import com.sparta.restplaceforj.dto.PostRequestDto;
@@ -12,7 +10,6 @@ import com.sparta.restplaceforj.dto.PostResponseDto;
 import com.sparta.restplaceforj.security.UserDetailsImpl;
 import com.sparta.restplaceforj.service.PostService;
 import jakarta.validation.Valid;
-import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -24,9 +21,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 /**
  * 여행 추천 글 api.
@@ -196,27 +191,6 @@ public class PostController {
         CommonResponse.<PostResponseDto>builder()
             .response(ResponseEnum.GET_POST)
             .data(postResponseDto)
-            .build()
-    );
-  }
-
-  /**
-   * 사진을 s3 업로드 api.
-   *
-   * @param images 저장할 파일
-   * @return ImageResponseDto : id, path, originalFileName, changedFiledName
-   * @throws IOException InputStream getInputStream() throws IOException
-   */
-  @PostMapping(value = "/posts/images")
-  public ResponseEntity<CommonResponse<ImageResponseDto>> createPostImage(
-      @RequestPart("images") MultipartFile images) throws IOException {
-
-    ImageResponseDto imageResponseDto = postService.createPostImage(images);
-
-    return ResponseEntity.ok(
-        CommonResponse.<ImageResponseDto>builder()
-            .response(ResponseEnum.CREATE_USER_PROFILE_IMAGE)
-            .data(imageResponseDto)
             .build()
     );
   }
